Rename Hitter2D to HitTester2D
